+++ examples/KPRCA_00111/cb_2/lib/stdint.h
@@ -13,8 +13,8 @@ typedef unsigned int uint32_t;
 typedef unsigned long long int uint64_t;
 typedef unsigned long long int uintmax_t;
 
-typedef int intptr_t;
-typedef unsigned int uintptr_t;
+typedef long int intptr_t;
+typedef unsigned long int uintptr_t;
 
 /* Minimum of signed integral types.  */
 #define INT8_MIN       (-128)
